backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.54
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==1.10.13
redis==5.0.1
anthropic==0.40.0
python-dotenv==1.0.0
httpx==0.25.2
sse-starlette==1.6.5
